+++ cyxbs-applications/test/src/jsMain/kotlin/CyxbsWebApp.js.kt
@@ -1,0 +1,11 @@
+import com.g985892345.provider.cyxbsmobile.cyxbsapplications.test.TestKtProviderInitializer
+
+/**
+ * .
+ *
+ * @author 985892345
+ * @date 2025/9/30
+ */
+actual fun initKtProvider() {
+  TestKtProviderInitializer.tryInitKtProvider()
+}
