--- cyxbs-applications/test/src/wasmJsMain/resources/index.html
@@ -1,26 +1,0 @@
-<!DOCTYPE html>
-<html lang="zh-cn">
-<head>
-  <meta charset="UTF-8">
-  <meta name="viewport" content="width=device-width, initial-scale=1.0">
-  <title>网上重邮</title>
-  <link type="text/css" rel="stylesheet" href="styles.css">
-  <script type="application/javascript" src="CyxbsApplicationsTest.js"></script>
-</head>
-<body>
-<div id="compose">
-</div>
-</body>
-<style>
-  body, html {
-    height: 100%; /* 确保整个页面高度 */
-    margin: 0; /* 移除默认边距 */
-    display: flex;
-    flex-direction: column;
-  }
-  #compose {
-    height: 100%;
-    width: 100%;
-  }
-</style>
-</html>
