+++ cyxbs-components/utils/src/jsMain/kotlin/com/cyxbs/components/utils/network/Network.js.kt
@@ -3,11 +3,6 @@ package com.cyxbs.components.utils.network
 import io.ktor.client.engine.HttpClientEngine
 import io.ktor.client.engine.js.Js
 
-/**
- * .
- *
- * @author 985892345
- * @date 2025/1/5
- */
 internal actual fun createHttpClientEngine(): HttpClientEngine = Js.create {
+
 }
