+++ cyxbs-components/utils/src/webMain/kotlin/com/cyxbs/components/utils/network/Network.web.kt
@@ -1,0 +1,12 @@
+package com.cyxbs.components.utils.network
+
+import io.ktor.client.HttpClientConfig
+
+/**
+ * .
+ *
+ * @author 985892345
+ * @date 2025/1/5
+ */
+internal actual fun HttpClientConfig<*>.platformConfigHttpClient() {
+}
