+++ gradle.properties
@@ -53,5 +53,5 @@ kotlin.mpp.applyDefaultHierarchyTemplate=false
 
 # 这里默认关闭，可以在 local.properties 覆盖打开配置
 cyxbs.multiplatform.ios=true
-cyxbs.multiplatform.wasm=false
+cyxbs.multiplatform.web=true
 cyxbs.multiplatform.desktop=true # 只有一个平台时 commonMain 可以调用安卓代码，所以这里建议至少开启一个
