+++ lib_common/src/main/java/com/mredrock/cyxbs/common/viewmodel/BaseViewModel.kt
@@ -3,7 +3,6 @@ package com.mredrock.cyxbs.common.viewmodel
 import androidx.annotation.CallSuper
 import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.ViewModel
-import com.mredrock.cyxbs.common.utils.extensions.longToast
 import com.mredrock.cyxbs.common.viewmodel.event.ProgressDialogEvent
 import com.mredrock.cyxbs.common.viewmodel.event.SingleLiveEvent
 import io.reactivex.rxjava3.disposables.Disposable
